Animatable

+++ 4-React-Native/nucampsite/components/FavoritesComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { FlatList, View, Text,StyleSheet,Alert } from 'react-native';
+import { FlatList, View, Text, StyleSheet, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { Loading } from './LoadingComponent';
@@ -7,6 +7,7 @@ import { baseUrl } from '../shared/baseUrl';
 import { SwipeRow } from 'react-native-swipe-list-view';
 import { TouchableOpacity } from 'react-native-gesture-handler';
 import { deleteFavorite } from '../redux/ActionCreators';
+import * as Animatable from 'react-native-animatable';
 
 const mapStateToProps = state => {
     return {
@@ -27,48 +28,48 @@ class Favorites extends Component {
         title: 'My Favorites'
     }
 
-    render(){
+    render() {
         const { navigate } = this.props.navigation;
-        const renderFavoriteItem = ({item}) => {
+        const renderFavoriteItem = ({ item }) => {
             return (
                 <SwipeRow rightOpenValue={-100} style={styles.swipeRow}>
-                   <View style={styles.deleteView}>
-                    <TouchableOpacity
-                        style={styles.deleteTouchable}
-                        onPress={() => 
-                            Alert.alert(
-                            'Delete Favorite',
-                            'Are you sure you wish to delete the favorite campsite ' +
-                             item.name +
-                            '?',
-                            [
-                                {
-                                    text: 'Cancel',
+                    <View style={styles.deleteView}>
+                        <TouchableOpacity
+                            style={styles.deleteTouchable}
+                            onPress={() =>
+                                Alert.alert(
+                                    'Delete Favorite',
+                                    'Are you sure you wish to delete the favorite campsite ' +
+                                    item.name +
+                                    '?',
+                                    [
+                                        {
+                                            text: 'Cancel',
                                             onPress: () => console.log(item.name + 'Not Deleted'),
                                             style: 'cancel'
-                                },
-                                {
-                                    text: 'OK',
+                                        },
+                                        {
+                                            text: 'OK',
                                             onPress: () => this.props.deleteFavorite(item.id)
-                                },
-                            ],
-                           
-                           { cancelable: false }
-                         )
-                        }
-                        >
-                        <Text style={styles.deleteText}>Delete</Text>
-                     </TouchableOpacity>
+                                        },
+                                    ],
 
-                  </View>
-                  <View>
-                     <ListItem
-                         title={item.name}
-                         subtitle={item.description}
-                         leftAvatar={{source: {uri: baseUrl + item.image}}}
-                         onPress={() => navigate('CampsiteInfo', {campsiteId: item.id})}
-                     />
-                  </View>
+                                    { cancelable: false }
+                                )
+                            }
+                        >
+                            <Text style={styles.deleteText}>Delete</Text>
+                        </TouchableOpacity>
+
+                    </View>
+                    <View>
+                        <ListItem
+                            title={item.name}
+                            subtitle={item.description}
+                            leftAvatar={{ source: { uri: baseUrl + item.image } }}
+                            onPress={() => navigate('CampsiteInfo', { campsiteId: item.id })}
+                        />
+                    </View>
                 </SwipeRow>
             );
         };
@@ -83,19 +84,21 @@ class Favorites extends Component {
                 </View>
             );
         }
-        return(
-            <FlatList
-                data={this.props.campsites.campsites.filter(
-                    campsite => this.props.favorites.includes(campsite.id)
-                )}
-                renderItem={renderFavoriteItem}
-                keyExtractor={item => item.id.toString()}
-            />
-    
-        );  
+        return (
+            <Animatable.View animation='fadeInRightBig' duration={2000}>
+                <FlatList
+                    data={this.props.campsites.campsites.filter(
+                        campsite => this.props.favorites.includes(campsite.id)
+                    )}
+                    renderItem={renderFavoriteItem}
+                    keyExtractor={item => item.id.toString()}
+                />
+            </Animatable.View>
+
+        );
 
     }
-    
+
 }
 const styles = StyleSheet.create({
     deleteView: {
@@ -117,4 +120,4 @@ const styles = StyleSheet.create({
         width: 100
     }
 });
-export default connect(mapStateToProps,mapDispatchToProps)(Favorites);
+export default connect(mapStateToProps, mapDispatchToProps)(Favorites);
